pointing in get_mem
