+++ base_agent/dialogue_objects/interpreter.py
@@ -198,13 +198,3 @@ class Interpreter(DialogueObject):
     def handle_otheraction(self, speaker, d) -> Tuple[Optional[str], Any]:
         self.finished = True
         return "I don't know how to do that yet", None
-
-    def append_new_task(self, cls, data=None):
-        # this is badly named, FIXME
-        # add a tick to avoid two tasks having same timestamp
-        self.memory.add_tick()
-        if data is None:
-            self.memory.task_stack_push(cls, chat_effect=True)
-        else:
-            task = cls(self.agent, data)
-            self.memory.task_stack_push(task, chat_effect=True)
