+++ craftassist/agent/dialogue_objects/get_memory_handler.py
@@ -14,7 +14,7 @@ from base_agent.dialogue_objects import (
 from .attribute_helper import MCAttributeInterpreter
 from .spatial_reasoning import ComputeLocations
 from base_agent.base_util import ErrorWithResponse
-from tasks import Build
+from tasks import Build, Point
 from ttad.generation_dialogues.generate_utils import prepend_a_an
 
 
@@ -40,6 +40,9 @@ class MCGetMemoryHandler(GetMemoryHandler):
             "specify_locations": ComputeLocations(),
         }
         self.subinterpret["attribute"] = MCAttributeInterpreter()
+        self.task_objects = {
+            "point": Point
+        }
 
     def handle_task_refobj_string(self, task, refobj_attr):
         if refobj_attr == "name":
