+++ locobot/agent/dialogue_objects/get_memory_handler.py
@@ -18,6 +18,7 @@ from base_agent.memory_nodes import MemoryNode, ReferenceObjectNode
 from base_agent.string_lists import ACTION_ING_MAPPING
 from ttad.generation_dialogues.generate_utils import prepend_a_an
 from copy import deepcopy
+from tasks import Point
 
 
 class LocoGetMemoryHandler(GetMemoryHandler):
@@ -41,6 +42,9 @@ class LocoGetMemoryHandler(GetMemoryHandler):
             "reference_locations": ReferenceLocationInterpreter(),
             "specify_locations": ComputeLocations(),
             "attribute": AttributeInterpreter(),
+        }
+        self.task_objects = {
+            "point": Point
         }
 
     def handle_task_refobj_string(self, task, refobj_attr):
